+++ remat/core/solvers/strategy_checkpoint_last.py
@@ -2,9 +2,9 @@ import numpy as np
 
 from remat.core.dfgraph import DFGraph
 from remat.core.schedule import ScheduledResult
-from remat.core.utils.solvers.common import SOLVER_DTYPE, solve_r_opt
+from remat.core.utils.solver_common import SOLVER_DTYPE, solve_r_opt
 from remat.core.enum_strategy import SolveStrategy
-from remat.core.utils.solvers.scheduler import schedule_from_rs
+from remat.core.utils.scheduler import schedule_from_rs
 from remat.core.utils.timer import Timer
