+++ remat/core/solvers/strategy_chen.py
@@ -2,9 +2,9 @@ import math
 
 from remat.core.dfgraph import DFGraph
 from remat.core.schedule import ScheduledResult
-from remat.core.utils.solvers.common import gen_s_matrix_fixed_checkpoints, solve_r_opt
+from remat.core.utils.solver_common import gen_s_matrix_fixed_checkpoints, solve_r_opt
 from remat.core.enum_strategy import SolveStrategy
-from remat.core.utils.solvers.scheduler import schedule_from_rs
+from remat.core.utils.scheduler import schedule_from_rs
 from remat.core.utils.timer import Timer
