+++ remat/core/solvers/strategy_griewank.py
@@ -9,8 +9,8 @@ import pandas as pd
 
 from remat.core.dfgraph import DFGraph
 from remat.core.schedule import ScheduledResult
-from remat.core.utils.solvers.common import solve_r_opt, setup_implied_s_backwards
-from remat.core.utils.solvers.scheduler import schedule_from_rs
+from remat.core.utils.solver_common import solve_r_opt, setup_implied_s_backwards
+from remat.core.utils.scheduler import schedule_from_rs
 from remat.core.enum_strategy import SolveStrategy
 from remat.core.utils.timer import Timer
