+++ remat/core/solvers/strategy_optimal_ilp.py
@@ -9,8 +9,8 @@ from gurobipy import GRB, Model, quicksum
 import remat.core
 from remat.core.dfgraph import DFGraph
 from remat.core.schedule import ScheduledResult, ILPAuxData
-from remat.core.utils.solvers.common import solve_r_opt
-from remat.core.utils.solvers.scheduler import schedule_from_rs
+from remat.core.utils.solver_common import solve_r_opt
+from remat.core.utils.scheduler import schedule_from_rs
 from remat.core.enum_strategy import SolveStrategy
 from remat.core.utils.timer import Timer
 
@@ -153,10 +153,10 @@ class ILPSolver:
         if self.m.solCount < 1:
             raise ValueError(f"Model status is {self.m.status} (not infeasible), but solCount is {self.m.solCount}")
 
-        Rout = np.zeros((T, T), dtype=remat.core.utils.solvers.common.SOLVER_DTYPE)
-        Sout = np.zeros((T, T), dtype=remat.core.utils.solvers.common.SOLVER_DTYPE)
-        Uout = np.zeros((T, T), dtype=remat.core.utils.solvers.common.SOLVER_DTYPE)
-        Free_Eout = np.zeros((T, len(self.g.edge_list)), dtype=remat.core.utils.solvers.common.SOLVER_DTYPE)
+        Rout = np.zeros((T, T), dtype=remat.core.utils.solver_common.SOLVER_DTYPE)
+        Sout = np.zeros((T, T), dtype=remat.core.utils.solver_common.SOLVER_DTYPE)
+        Uout = np.zeros((T, T), dtype=remat.core.utils.solver_common.SOLVER_DTYPE)
+        Free_Eout = np.zeros((T, len(self.g.edge_list)), dtype=remat.core.utils.solver_common.SOLVER_DTYPE)
         try:
             for t in range(T):
                 for i in range(T):
