+++ remat/core/utils/solver_common.py
